Unity version can now be parsed from ProjectVersion.txt!!!
This is done without a Target, so the generated Reference paths will still be visible in the VS Solution Explorer! Also renamed UnityVersion to UnityEditorVersion, to better fit the values in ProjectVersion.txt (and updated usage of the property in test projects).

This parsing is not performed when the version is set explicitly (e.g., for referencing Unity assemblies without an associated project). If a UnityProjectPath is not provided, then the version instead becomes "VERSION_OR_PROJECT_NOT_SET", to instruct users on what's missing.