Package signing is now optional in publish script
Now that I'm publishing this as an individual not an organization, having a signing cert is a lot less meaningful. Not to mention they're expensive and I don't want my contact info floating around in a cert...probably just gonna stop signing. v2.1.0 is already unsigned cause I wanted to get it out the door. We'll see how things go for future versions...